further gem screwing
